refactor(dashboard): change AuthorizationInterceptor to an interface
